refactor(platform): Standardize file naming with pal_ prefix
- Renames all files in platform/ to follow consistent pal_ prefix.
- Improves discoverability and enforces uniform naming convention.
