+++ platform/pal_baremetal/FVP/include/pal_image_def.h
