+++ platform/pal_uefi/include/pal_override.h
